backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pywebpush
python-dotenv
